requirements.txt
pyalsaaudio==0.7
